requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-amazon==7.2.1
requests==2.31.0